move CompressionAlgorithm and GzipLevel
