+++ nexus/reconfigurator/planning/src/blueprint_editor/sled_editor/datasets.rs
@@ -10,12 +10,12 @@ use illumos_utils::zpool::ZpoolName;
 use nexus_types::deployment::BlueprintDatasetConfig;
 use nexus_types::deployment::BlueprintDatasetDisposition;
 use omicron_common::api::external::ByteCount;
-use omicron_common::disk::CompressionAlgorithm;
 use omicron_common::disk::DatasetKind;
 use omicron_common::disk::DatasetName;
-use omicron_common::disk::GzipLevel;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::ZpoolUuid;
+use sled_agent_types::disk::CompressionAlgorithm;
+use sled_agent_types::disk::GzipLevel;
 use std::collections::BTreeMap;
 use std::collections::btree_map::Entry;
 use std::net::SocketAddrV6;
