+++ nexus/reconfigurator/rendezvous/src/debug_dataset.rs
@@ -144,10 +144,10 @@ mod tests {
     use nexus_db_queries::db::pub_test_utils::TestDatabase;
     use nexus_db_queries::db::queries::ALLOW_FULL_TABLE_SCAN_SQL;
     use nexus_types::inventory::ZpoolName;
-    use omicron_common::disk::CompressionAlgorithm;
     use omicron_test_utils::dev;
     use proptest::prelude::*;
     use proptest::proptest;
+    use sled_agent_types::disk::CompressionAlgorithm;
 
     async fn proptest_do_prep(
         opctx: &OpContext,
