+++ nexus/types/src/deployment/blueprint_diff.rs
@@ -24,9 +24,10 @@ use super::{
 };
 use daft::{Diffable, Leaf};
 use omicron_common::api::external::ByteCount;
-use omicron_common::disk::{CompressionAlgorithm, DatasetName};
+use omicron_common::disk::DatasetName;
 use omicron_uuid_kinds::SledUuid;
 use omicron_uuid_kinds::{DatasetUuid, OmicronZoneUuid, PhysicalDiskUuid};
+use sled_agent_types_versions::latest::disk::CompressionAlgorithm;
 use sled_agent_types_versions::latest::inventory::ZoneKind;
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::{self, Write as _};
