+++ sled-agent/rack-setup/src/plan/service.rs
@@ -43,7 +43,7 @@ use omicron_common::api::internal::shared::{
 use omicron_common::backoff::{
     BackoffError, retry_notify_ext, retry_policy_internal_service_aggressive,
 };
-use omicron_common::disk::{CompressionAlgorithm, DatasetKind, DatasetName};
+use omicron_common::disk::{DatasetKind, DatasetName};
 use omicron_common::policy::{
     BOUNDARY_NTP_REDUNDANCY, COCKROACHDB_REDUNDANCY,
     CRUCIBLE_PANTRY_REDUNDANCY, INTERNAL_DNS_REDUNDANCY, NEXUS_REDUNDANCY,
@@ -60,6 +60,7 @@ use serde::{Deserialize, Serialize};
 use sled_agent_client::{
     Client as SledAgentClient, Error as SledAgentError, types as SledAgentTypes,
 };
+use sled_agent_types::disk::CompressionAlgorithm;
 use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::disk::DiskVariant;
 use sled_agent_types::disk::SharedDatasetConfig;
