+++ sled-agent/src/support_bundle/storage.rs
@@ -16,7 +16,6 @@ use futures::StreamExt;
 use illumos_utils::zfs::DatasetProperties;
 use illumos_utils::zfs::DestroyDatasetError;
 use illumos_utils::zfs::DestroyDatasetErrorVariant;
-use omicron_common::disk::CompressionAlgorithm;
 use omicron_common::disk::DatasetName;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::SupportBundleUuid;
@@ -31,6 +30,7 @@ use sled_agent_config_reconciler::NestedDatasetDestroyError;
 use sled_agent_config_reconciler::NestedDatasetEnsureError;
 use sled_agent_config_reconciler::NestedDatasetListError;
 use sled_agent_config_reconciler::NestedDatasetMountError;
+use sled_agent_types::disk::CompressionAlgorithm;
 use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::disk::SharedDatasetConfig;
 use sled_agent_types::support_bundle::BUNDLE_FILE_NAME;
