+++ sled-agent/types/versions/src/impls/disk.rs
@@ -2,8 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use crate::latest::disk::CompressionAlgorithm;
 use crate::latest::disk::DiskStateRequested;
 use crate::latest::disk::DiskVariant;
+use crate::latest::disk::GzipLevel;
 use crate::latest::disk::M2Slot;
 use anyhow::bail;
 use omicron_common::zpool_name::ZpoolKind;
@@ -94,5 +96,71 @@ impl From<ZpoolKind> for DiskVariant {
             ZpoolKind::External => DiskVariant::U2,
             ZpoolKind::Internal => DiskVariant::M2,
         }
+    }
+}
+
+// Fastest compression level
+const GZIP_LEVEL_MIN: u8 = 1;
+
+// Best compression ratio
+const GZIP_LEVEL_MAX: u8 = 9;
+
+impl GzipLevel {
+    pub const fn new<const N: u8>() -> Self {
+        assert!(N >= GZIP_LEVEL_MIN, "Compression level too small");
+        assert!(N <= GZIP_LEVEL_MAX, "Compression level too large");
+        Self(N)
+    }
+}
+
+impl FromStr for GzipLevel {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let level = s.parse::<u8>()?;
+        if level < GZIP_LEVEL_MIN || level > GZIP_LEVEL_MAX {
+            bail!("Invalid gzip compression level: {level}");
+        }
+        Ok(Self(level))
+    }
+}
+
+/// These match the arguments which can be passed to "zfs set compression=..."
+impl fmt::Display for CompressionAlgorithm {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let s = match self {
+            Self::On => "on",
+            Self::Off => "off",
+            Self::Gzip => "gzip",
+            Self::GzipN { level } => {
+                return write!(f, "gzip-{}", level.0);
+            }
+            Self::Lz4 => "lz4",
+            Self::Lzjb => "lzjb",
+            Self::Zle => "zle",
+        };
+        write!(f, "{}", s)
+    }
+}
+
+impl FromStr for CompressionAlgorithm {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let c = match s {
+            "on" => Self::On,
+            "" | "off" => Self::Off,
+            "gzip" => Self::Gzip,
+            "lz4" => Self::Lz4,
+            "lzjb" => Self::Lzjb,
+            "zle" => Self::Zle,
+            _ => {
+                let Some(suffix) = s.strip_prefix("gzip-") else {
+                    bail!("Unknown compression algorithm {s}");
+                };
+                Self::GzipN { level: suffix.parse()? }
+            }
+        };
+        Ok(c)
     }
 }
